test: 동일한 시간대에 이미 예매한 내역이 있으면 예외가 발생한다

+++ src/main/kotlin/domain/User.kt
@@ -1,5 +1,5 @@
 package domain
-class User(val id: Long) {
+class User(val id: Int) {
     var point: Long = 1000
 
     init {
